# services/notificationservice/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Notification Service API
  description: APi for managing user notitications in LMS
  version: 1.0.0

servers:
  - url: http://localhost:8089/api/v1/
    description: Development server

paths:
  /notifications/users/me:
    get:
      summary: Get my notifications
      description: Retrieves list of all notifications for the currently authenticated user
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /notifications/unread:
    get:
      summary: Get my all unread notifications
      description: Retrieves list of all notifications which is unread for the currently authenticated user
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of notifications which 'read' field equals false
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /notifications/count:
    get:
      summary: Get count of unread notifications
      description: Return a count of unread messages
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Map of {"unreadCount", count}
          content:
            application/json:
              schema:
                type: number
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /notifications/{id}/read:
    put:
      summary: Mark a notification as read
      description: Marks a specific notification as read for the authenticated user.
      tags:
        - Notifications
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the notification to mark as read.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDto'
        '403':
          description: Forbidden (if user tries to read another user's notification)
        '404':
          description: Notification not found
  /notifications/read-all:
    put:
      summary: Mark all notifications as read
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDto'
        '403':
          description: Forbidden (if user tries to read another user's notification)
        '401':
          description: Unauthorized
  /notifications/{id}:
    delete:
      summary: Delete notification
      description: Delete notification by id
      tags:
        - Notifications
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Id of notification what user want to delete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Void
        '403':
          description: Forbidden (if user tries to delete another user's notification)
        '401':
          description: Unauthorized
  /notifications/:
    post:
      summary: Create notification
      description: Create notification can only admin
      tags:
        - Notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Returns created notification
        '401':
          description: Unauthorized
        '403':
          description: You are not an admin
  /notifications/users/{userId}:
    get:
      summary: Get user notifications
      description: Get list of all user notifications, require admin
      tags:
        - Notifications
      security:
        - bearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: Id of notification what user want to delete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of notifications
        '401':
          description: Unauthorized
        '403':
          description: You are not an admin

components:
  schemas:
    NotificationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the notification.
        type:
          type: string
          enum: [WELCOME, COURSE_ENROLLMENT, NEW_MODULE, ADMIN_MESSAGE]
          description: The type of the notification.
        title:
          type: string
          description: The title of the notification.
          example: "Welcome to the Platform!"
        message:
          type: string
          description: The main content of the notification.
        isRead:
          type: boolean
          description: Flag indicating if the notification has been read.
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the notification was created.

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT